Refactor promote functions
